move common connecting functions to separate file
